refactor: testing refactoring
